requirements.txt
streamlit==1.15.2
pandas==1.5.2
streamlit-echarts==0.4.0
Pillow==9.0.1